formatting flight results
